% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_network.R
\name{check_network}
\alias{check_network}
\title{Check Graph Network}
\usage{
check_network(graph)
}
\arguments{
\item{graph}{igraph}
}
\description{
\code{check_network} returns processed graph.
}
\details{
This is a pre-processing function that turns raw input into directed
networks with edge weights.
}
