% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{create_macro}
\alias{create_macro}
\title{create_macro}
\usage{
create_macro(graph, mapping, macro_types, ...)
}
\arguments{
\item{graph}{igraph}

\item{mapping}{List mapping from micro to macro nodes.}

\item{macro_types}{List of node distribution types.}

\item{...}{Passed arguments.}
}
\value{
Directed igraph graph object corresponding to a coarse-grained network
according to the \code{mapping} of micro nodes onto macro nodes, given by \code{mapping}.
}
\description{
Coarse-grains a network according to the specified macro_mapping and
the types of macros that each macro is associated with.
}
