% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einet.R
\docType{package}
\name{einet}
\alias{einet}
\alias{einet-package}
\title{einet: Uncertainty and causal emergence in complex networks.}
\description{
for calculating effective information in networks.
This can then be used to search for macroscale representations of
a network such that the coarse grained representation has more effective
information than the microscale, a phenomenon known as causal emergence.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/travisbyrum/einet}
  \item Report bugs at \url{https://github.com/travisbyrum/einet/issues}
}

}
\author{
\strong{Maintainer}: Travis Byrum \email{tbyrum@terpmail.umd.edu}

Authors:
\itemize{
  \item Anshuman Swain \email{answain@terpmail.umd.edu}
  \item Brennan Klein \email{klein.br@northeastern.edu}
  \item William Fagan \email{bfagan@umd.edu}
}

}
