% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{karate}
\alias{karate}
\title{Zachary's karate club}
\format{Igraph object with 78 edges.}
\source{
\url{http://www-personal.umich.edu/~mejn/netdata/}
}
\usage{
karate
}
\description{
Social network data of university karate club.  Used for causal emergence benchmarking and testing.
}
\keyword{datasets}
