% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{select_macro}
\alias{select_macro}
\title{select_macro}
\usage{
select_macro(graph, macro, mapping, macro_types, ...)
}
\arguments{
\item{graph}{igraph object of micro network}

\item{macro}{numeric of possible macro node}

\item{mapping}{List mapping from micro to macro nodes.}

\item{macro_types}{List of node distribution types.}

\item{...}{Passed arguments.}
}
\value{
List with igraph graph object corresponding to a coarse-grained network
and updated node types.
}
\description{
Given a potential mapping of the network and a new macro node,
this function assigns a new macro node to the candidate node optimized for
maximal accuracy.
}
