% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_emergence.R
\name{stationary}
\alias{stationary}
\title{Stationary Distribution}
\usage{
stationary(graph, zero_cutoff = 1e-10)
}
\arguments{
\item{graph}{igraph or matrix object.}

\item{zero_cutoff}{Numeric threshold for zero value.}
}
\value{
A numeric \code{vector} corresponding to stationary distribution.
}
\description{
Gives a stationary probability vector of a given network.
}
