% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_emergence.R
\name{update_blanket}
\alias{update_blanket}
\title{Update Markov Blanket}
\usage{
update_blanket(blanket, removal = NULL)
}
\arguments{
\item{blanket}{List of previous markov blanket.}

\item{removal}{Numeric vector for node removal.}
}
\description{
Update Markov Blanket
}
