% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_emergence.R
\name{causal_emergence}
\alias{causal_emergence}
\title{Causal Emergence}
\usage{
causal_emergence(x, ...)
}
\arguments{
\item{x}{igraph or matrix object.}

\item{...}{Span, and threshold parameters}
}
\description{
Calculates casual emergence.
}
