% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_emergence.R
\name{mb}
\alias{mb}
\title{Create Markov Blanket}
\usage{
mb(graph, nodes = igraph::V(graph))
}
\arguments{
\item{graph}{igraph or matrix object.}

\item{nodes}{Numeric vector of vertices.}
}
\description{
Create Markov Blanket
}
